// src/types/types.ts
/* eslint-disable no-unused-vars */
export enum WalletStatus {
    success = 'success',
    error = 'error',
    warning = 'warning',
    info = 'info',
}

export enum CalculatorSpace {
    small  =  "10000",
    middle =  "50000",
    large  =  "100000"
}

export enum SliderImageIndex {
    startIndex = 0,
    zero = 0,
    one = 1,
    two = 2,
    tree = 3,
    four = 4,
    five = 5,
    six  = 6
}
export enum EmptyLength  {
    zero = 0,
    one =  1
}